backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
psycopg2-binary
gunicorn
google-generativeai
python-docx
whitenoise
Pillow
